fixed some compiler warnings fixed "map/set iterator not incrementable" bug while change key binding

+++ src/game/bullets/Flame.cpp
@@ -12,8 +12,8 @@ Flame::Flame(float x, float y, Texture* tex) :
 	m_img = new StaticObject(x, y, 128, 128, tex, false);
 	m_img->RMask = 1.0;
 	m_img->GMask = (float) (rand() % 50) / 100 + 0.4;
-	m_img->BMask = 0.3;
-	m_img->Scale = 0.001;
+	m_img->BMask = 0.3f;
+	m_img->Scale = 0.001f;
 }
 
 void Flame::process(int deltaTime) {
@@ -27,20 +27,20 @@ void Flame::process(int deltaTime) {
 
 	const float relSpeed = (1.0 - dist) * Speed;
 
-	X -= cos((Angle + 90) * M_PI / 180) * deltaTime * relSpeed;
-	Y -= sin((Angle + 90) * M_PI / 180) * deltaTime * relSpeed;
+	X -= (float)cos((Angle + 90) * M_PI / 180) * deltaTime * relSpeed;
+	Y -= (float)sin((Angle + 90) * M_PI / 180) * deltaTime * relSpeed;
 
 	m_img->X = X;
 	m_img->Y = Y;
 
 	m_range += relSpeed * deltaTime;
 
-	m_img->Scale = 2.5 * dist;
+	m_img->Scale = 2.5f * dist;
 
 	if (dist < 0.5)
 		m_img->AMask = dist;
 	else
-		m_img->AMask = 1.0 - dist;
+		m_img->AMask = 1.0f - dist;
 }
 
 void Flame::draw() {
